+++ lib/pages/detection/index.dart
